src/agents/real_time_agent.py
"""
NeoMate AI Real-Time Agent Module

This module implements the RealTimeAgent class for handling queries that require
current information from the web. It integrates web search capabilities to provide
up-to-date responses for time-sensitive queries.

Features:
- Web search integration for current information
- Query optimization for search engines
- Result summarization and formatting
- Error handling for search failures
- Context-aware search queries

Usage:
    from src.agents.real_time_agent import RealTimeAgent

    agent = RealTimeAgent()
    response = agent.handle_query("What is today's weather?")
"""

import requests
from typing import Tuple, Any, Dict
from src.utils.logger import logger


class RealTimeAgent:
    """
    Agent for handling real-time queries requiring web search.

    Provides current information by searching the web and summarizing
    results for queries about news, weather, current events, etc.
    """

    def __init__(self):
        self.agent_type = "real_time"
        # Placeholder for search API configuration
        self.search_api_key = None  # Would be loaded from config
        self.search_engine_url = "https://www.googleapis.com/customsearch/v1"

    def handle_query(self, query: str) -> str:
        """
        Handle a real-time query by performing web search.

        Args:
            query: User query requiring current information.

        Returns:
            Summarized search results string.
        """
        try:
            # Optimize query for search
            search_query = self._optimize_query(query)

            # Perform search (placeholder implementation)
            results = self._perform_search(search_query)

            # Summarize and format results
            response = self._format_results(query, results)

            logger.info("RealTimeAgent successfully handled query")
            return response

        except Exception as e:
            logger.error(f"RealTimeAgent failed to handle query: {e}")
            return "I'm sorry, I couldn't find current information for your query. Please try again."

    def execute(self, tool: str, args: dict) -> Tuple[bool, str]:
        """
        Execute a tool action.

        Args:
            tool: Tool name ('web_search' or 'gather_information')
            args: Tool arguments

        Returns:
            Tuple of (success: bool, result: str)
        """
        if tool == 'web_search':
            query = args.get('query', '')
            response = self.handle_query(query)
            return True, response
        elif tool == 'gather_information':
            info_type = args.get('info_type', '')
            query = f"Current information about {info_type}"
            response = self.handle_query(query)
            return True, response
        else:
            return False, f"RealTimeAgent cannot execute tool: {tool}"

    def _optimize_query(self, query: str) -> str:
        """
        Optimize query for better search results.

        Args:
            query: Original user query

        Returns:
            Optimized search query
        """
        # Add current date/time context for time-sensitive queries
        import datetime
        current_year = datetime.datetime.now().year

        # Add time-sensitive keywords if needed
        if any(word in query.lower() for word in ['current', 'today', 'now', 'latest']):
            query = f"{query} {current_year}"

        return query

    def _perform_search(self, query: str) -> list:
        """
        Perform web search (placeholder implementation).

        Args:
            query: Search query string

        Returns:
            List of search results (placeholder)
        """
        # Placeholder for actual search implementation
        # In a real implementation, this would call a search API
        # For now, return mock results
        return [
            {
                "title": f"Search result for: {query}",
                "snippet": f"This is a placeholder result for the query '{query}'. In a real implementation, this would contain actual web search results.",
                "url": "https://example.com"
            }
        ]

    def _format_results(self, original_query: str, results: list) -> str:
        """
        Format search results into a readable response.

        Args:
            original_query: Original user query
            results: List of search results

        Returns:
            Formatted response string
        """
        if not results:
            return f"I couldn't find information about '{original_query}'."

        # Format the first result (in real implementation, might summarize multiple)
        result = results[0]
        response = f"Based on current information: {result['snippet']}"

        return response
